Refactor: fetch join과 chatRoom에 currentMember 컬럼을 추가하여 N+1문제 해결
채팅방에 신규 멤버 참가 시 currentMember증가시키는 메서드 구현

Resolves: #153

+++ src/main/java/com/example/eat_together/domain/chat/dto/ChatRoomDto.java
@@ -16,9 +16,9 @@ public class ChatRoomDto {
     private FoodType foodType;
     private Integer maxMember;
     private ChatGroupStatus chatGroupStatus;
-    private Integer currentMemberCount;
+    private Long currentMemberCount;
 
-    public static ChatRoomDto of(Long roomId, String title, String description, FoodType foodType, Integer maxMember, ChatGroupStatus chatGroupStatus, Integer currentMemberCount) {
+    public static ChatRoomDto of(Long roomId, String title, String description, FoodType foodType, Integer maxMember, ChatGroupStatus chatGroupStatus, Long currentMemberCount) {
         ChatRoomDto chatRoomDto = new ChatRoomDto();
         chatRoomDto.roomId = roomId;
         chatRoomDto.title = title;
